filter_screened_poisson: Remove license-conflicting PLY file lib.
Modifications ported over from the copy of this software found in Geogram:
https://github.com/alicevision/geogram/commit/4a55fcc041a398f337fbc536ac7aab9c176cebf0
