+++ src/meshlabplugins/filter_screened_poisson/filter_screened_poisson.pro
@@ -10,7 +10,6 @@ HEADERS += \
 SOURCES += \
     filter_screened_poisson.cpp \
     Src/MarchingCubes.cpp \
-    Src/PlyFile.cpp \
     #Src/CmdLineParser.cpp \
     Src/Factor.cpp \
     Src/Geometry.cpp
