+++ src/templates/filter_screened_poisson.cmake
@@ -3,7 +3,6 @@
 {% block sources %}
 filter_screened_poisson.cpp
 Src/MarchingCubes.cpp
-Src/PlyFile.cpp
 # Src/CmdLineParser.cpp
 Src/Factor.cpp
 Src/Geometry.cpp
